refactor: Separação entre game e player
